Frontend: Replace evenChunkReader with a faster solution
Before evenChunkReader was making a really expensive copy that was
slowing down the encryption and decryption by a lot.

Now everything is copy-free and there's way less time between the
encryption operations.

After some profiling I set the CHUNK_SIZE to 2MB: this is the highest
amount I was able to set it to before the encryption task become longer
than 50ms (which, according to web.dev, should be avoided).